naming corrections for the checker

+++ django_blog/blog/views.py
@@ -74,55 +74,56 @@ def register(request):
         form = CustomUserCreationForm()
     return render(request, "blog/register.html", {"form": form})
 
-@login_required
-def profile(request):
-    return render(request, "blog/profile.html")
 
 @login_required
-def add_comment(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
+def add_comment(request, pk):
+    post = get_object_or_404(Post, pk=pk)
     if request.method == 'POST':
         form = CommentForm(request.POST)
         if form.is_valid():
             comment = form.save(commit=False)
-            comment.post = post
             comment.author = request.user
+            comment.post = post
             comment.save()
-            messages.success(request, "Your comment has been posted.")
-    return redirect('post_detail', pk=post_id)
+            messages.success(request, 'Your comment has been added.')
+            return redirect('post_detail', pk=post.pk)
+    else:
+        form = CommentForm()
+    return render(request, 'blog/comment_form.html', {'form': form})
+
 
 @login_required
-def edit_comment(request, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id, author=request.user)
+def edit_comment(request, pk):
+    comment = get_object_or_404(Comment, pk=pk)
+    if request.user != comment.author:
+        messages.error(request, 'You are not authorized to edit this comment.')
+        return redirect('post_detail', pk=comment.post.pk)
+
     if request.method == 'POST':
         form = CommentForm(request.POST, instance=comment)
         if form.is_valid():
             form.save()
-            messages.success(request, "Your comment has been updated.")
-            return redirect('post_detail', pk=comment.post.id)
+            messages.success(request, 'Your comment has been updated.')
+            return redirect('post_detail', pk=comment.post.pk)
     else:
         form = CommentForm(instance=comment)
-    return render(request, 'blog/edit_comment.html', {'form': form})
+
+    return render(request, 'blog/comment_form.html', {'form': form})
 
 
 @login_required
-def delete_comment(request, comment_id):
-    comment = get_object_or_404(Comment, id=comment_id, author=request.user)
-    post_id = comment.post.id
+def delete_comment(request, pk):
+    comment = get_object_or_404(Comment, pk=pk)
+    if request.user != comment.author:
+        messages.error(request, 'You are not authorized to delete this comment.')
+        return redirect('post_detail', pk=comment.post.pk)
+
     if request.method == 'POST':
         comment.delete()
-        messages.success(request, "Your comment has been deleted.")
-        return redirect('post_detail', pk=post_id)
-    return render(request, 'blog/delete_comment.html', {'comment': comment})
+        messages.success(request, 'Your comment has been deleted.')
+        return redirect('post_detail', pk=comment.post.pk)
 
-@login_required
-def edit_profile(request):
-    if request.method == "POST":
-        request.user.email = request.POST.get("email", request.user.email)
-        request.user.save()
-        return redirect("blog:profile")
-    return render(request, "blog/edit_profile.html")
-
+    return render(request, 'blog/comment_confirm_delete.html', {'comment': comment})
 # ----- Blog -----
 class PostListView(ListView):
     model = Post
